+++ Main/Adapters/IAgentAdapter.cs
@@ -23,7 +23,6 @@ public interface IAgentAdapter
 
     ActionIndexCache GetCurrentAction(int actionChannelNo);
     float GetCurrentActionProgress(int actionChannelNo);
-    bool IsWarg();
     bool IsHorse();
     bool IsCamel();
     bool IsActive();